Generate JSON From folder structure
This tool now generates a json file by searching through a directory for .html files and adding them to the json file.

The file can be saved from the tool.
For each .html file, the tool looks for a .txt file, and if found, populates the description variable by a description that should be written between two '*.desc' tags. For example:

*.desc This is my description! *.desc

-To do:
Add tags
Change from Blog to Page